[dagster-fivetran] reorganize Fivetran test suite
