+++ python_modules/libraries/dagster-fivetran/dagster_fivetran_tests/deprecated/test_asset_defs.py
@@ -10,7 +10,7 @@ from dagster_fivetran.resources import (
     FIVETRAN_CONNECTOR_PATH,
 )
 
-from dagster_fivetran_tests.utils import (
+from dagster_fivetran_tests.deprecated.utils import (
     DEFAULT_CONNECTOR_ID,
     get_sample_columns_response,
     get_sample_connector_response,
