+++ python_modules/libraries/dagster-fivetran/dagster_fivetran_tests/deprecated/test_ops.py
@@ -9,7 +9,7 @@ from dagster_fivetran.resources import (
     FIVETRAN_CONNECTOR_PATH,
 )
 
-from dagster_fivetran_tests.utils import (
+from dagster_fivetran_tests.deprecated.utils import (
     DEFAULT_CONNECTOR_ID,
     get_complex_sample_connector_schema_config,
     get_sample_connector_response,
