+++ python_modules/libraries/dagster-fivetran/dagster_fivetran_tests/test_asset_specs.py
@@ -12,7 +12,7 @@ from dagster_fivetran import (
 from dagster_fivetran.asset_defs import build_fivetran_assets_definitions
 from dagster_fivetran.translator import FivetranMetadataSet
 
-from dagster_fivetran_tests.beta.conftest import (
+from dagster_fivetran_tests.conftest import (
     TEST_ACCOUNT_ID,
     TEST_ANOTHER_ACCOUNT_ID,
     TEST_API_KEY,
